update côté front suite aux modifications des requêtes SQL, les participants sont maintenant affichés, et le bouton rejoindre est fonctionnel pour alimenter la table participants
