+++ frontend/src/components/LobbiesGameList.jsx
@@ -1,21 +1,9 @@
 /* eslint-disable eqeqeq */
-import { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import axios from "axios";
-import CurrentUserContext from "../contexts/CurrentUserContext";
 import MiniCard from "./MiniCard";
 
 export default function LobbiesGameList({ displayLobbies, trainNumberFilter }) {
-  const { currentUser } = useContext(CurrentUserContext);
   const navigate = useNavigate();
-
-  const handleSubmit = () => {
-    axios
-      .post(`${import.meta.env.VITE_BACKEND_URL}/participants`, {
-        user_id: currentUser.sub,
-      })
-      .catch((error) => console.error(error));
-  };
 
   return (
     <div>
@@ -31,22 +19,6 @@ export default function LobbiesGameList({ displayLobbies, trainNumberFilter }) {
             .map((lobby) => (
               <div className="mini-card" key={lobby.id}>
                 <MiniCard lobby={lobby} />
-                <div className="btn-container">
-                  <button
-                    type="button"
-                    className="details-btn"
-                    onClick={() => navigate(`/lobbies/${lobby.id}`)}
-                  >
-                    détails
-                  </button>
-                  <button
-                    type="button"
-                    className="join-btn"
-                    onClick={handleSubmit}
-                  >
-                    rejoindre
-                  </button>
-                </div>
               </div>
             ))}
       </div>
